// src/models/Card.ts
enum Card {
    c2 = "2c",
    c3 = "3c",
    c4 = "4c",
    c5 = "5c",
    c6 = "6c",
    c7 = "7c",
    c8 = "8c",
    c9 = "9c",
    cT = "Tc",
    cJ = "Jc",
    cQ = "Qc",
    cK = "Kc",
    cA = "Ac",

    d2 = "2d",
    d3 = "3d",
    d4 = "4d",
    d5 = "5d",
    d6 = "6d",
    d7 = "7d",
    d8 = "8d",
    d9 = "9d",
    dT = "Td",
    dJ = "Jd",
    dQ = "Qd",
    dK = "Kd",
    dA = "Ad",

    h2 = "2h",
    h3 = "3h",
    h4 = "4h",
    h5 = "5h",
    h6 = "6h",
    h7 = "7h",
    h8 = "8h",
    h9 = "9h",
    hT = "Th",
    hJ = "Jh",
    hQ = "Qh",
    hK = "Kh",
    hA = "Ah",

    s2 = "2s",
    s3 = "3s",
    s4 = "4s",
    s5 = "5s",
    s6 = "6s",
    s7 = "7s",
    s8 = "8s",
    s9 = "9s",
    sT = "Ts",
    sJ = "Js",
    sQ = "Qs",
    sK = "Ks",
    sA = "As",
    error = "error"
}
export default Card;